Rewritten some major idea's of the actor api. First of all, actors run on scheduled services now and basic support within actors is written for scheduled tasks that could be implemented in any kind of event/task system. Furthermore, I removed generics for the sender in messages, as in most cases all you need is to reply which can be handled within the message and only requires the actor object not specifically casted to its real implementation. In cases you need the sender reference it's still available.
I've done some major refactoring as well, method were removed, added again and they still subject to change, generics have been changed.

As for the MMO server implementation, I've written a start to player updating and a proper region system with a fancy loading mechanism, this will need some finishing work and is untested. Player updating however should be feasible quickly from its current state, the required logic is there all it needs is some glue.

From here on I will first add some proper documentation to the actor api as it seems that the current code in the actor api will remain more or less the same, changes in the future will most likely be in extension of features.